// lib/migration/ledger.generated.ts
export const queryMigrations: Array<{
  param: string;
  value: string;
  action: "301" | "410";
  destination: string | null;
}> = [
  {
    "param": "page_id",
    "value": "4958",
    "action": "301",
    "destination": "/pages/beginners-guide-to-bjj-signup"
  },
  {
    "param": "page_id",
    "value": "5157",
    "action": "301",
    "destination": "/pages/beginners-guide-to-bjj-signup"
  },
  {
    "param": "p",
    "value": "4942",
    "action": "410",
    "destination": null
  },
  {
    "param": "p",
    "value": "5110",
    "action": "410",
    "destination": null
  },
  {
    "param": "p",
    "value": "5030",
    "action": "410",
    "destination": null
  },
  {
    "param": "p",
    "value": "5246",
    "action": "410",
    "destination": null
  },
  {
    "param": "p",
    "value": "5257",
    "action": "410",
    "destination": null
  }
];